requirements.txt
absl-py==2.1.0
ale-py==0.8.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
deepdiff==7.0.1
dm-env==1.6
dm-tree==0.1.8
enum34==1.1.10
Farama-Notifications==0.0.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mock==5.1.0
mpmath==1.3.0
mpyq==0.2.5
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
portpicker==1.6.0
protobuf==3.20.0
psutil==6.0.0
pygame==2.6.0
pyparsing==3.1.2
PySC2==4.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
s2clientprotocol==5.0.12.91115.0
s2protocol==5.0.13.92440.3
scipy==1.14.0
Shimmy==0.2.1
six==1.16.0
sk-video==1.1.10
stable_baselines3==2.3.2
sympy==1.13.1
torch==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
